Update index.d.ts
There is an option to override the default identifiers but it is not available in the typescript config.
Also, the `composite` key in the index configuration is optional when one uses the `template`. 

TLDR: I am migrating an existing OneTable project to electroDb and I have to support things done slightly differently. None-the-less the types provided in this file right now are wrong