Fixes Issue 366 (#376)
* Initial fix, prior to understanding logical issue with `update`, `patch`, and `upsert`.

The operations `update`, `patch`, and `upsert` will need to change their current behavior when encountering a `false` value from `condition`. In that case, these operations will actually need to `delete` the `pk` and `sk` for that index.

* Solidifies new `condition` logic

The condition callback will be invoked only when a composite attribute associated with an index is set via an update, patch, or upsert. [existing behavior]
The condition callback is provided the attributes being set on that particular operation, including the item's identifying composite attributes. [existing behavior]
If the condition callback returns true, ElectroDB will attempt to create the index and all of its associated keys. If an index cannot be created because an update operation only has enough context for a partial key, ElectroDB will throw. [the original issue here, fixed]
If the condition callback returns false, the index and all of its associated keys will be removed from the item. [new behavior]
Item #1 above is the key to solving the issue you bring up in your first comment, and it's actually what we do currently. This means that condition would only be called when an index must be recalculated. furthermore, as described in #3, ElectroDB will actually throw if your update operation (set and remove) lacks a full composite context and would result in a "partial" key. This would mean that all * -> true transitions are already validated to have all the composite parts necessary to recreate the complete index already.

* Checkpoint commit
Checkpointing initial pass at new condition tests, tests not passing.

* All current tests working

* Clean up and test fix

* Clean up and test fix

* Clean up and test fix

* Clean up and test fix

* Clean up and test fix

* Clean up and adds new test cases

* adds new test case

* Adds changelog documentation

* Fixes test that used dynamic datetime

* Adds additional tests